+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log for Matrix
 
+## 2026-07-23
+- Add Docker environment
+
 ## 2026-07-22
 - Add spring-ai-starter-mcp-client and can take Claude Desktop JSON format.